requirements.txt
google-generativeai
PyMuPDF
python-dotenv
tqdm
difflib
pyttsx3
soundfile
numpy
pydub
